backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-filter>=24.1
Pillow>=10.2
stripe>=8.0
paypalrestsdk>=1.13
openai>=1.12
pgvector>=0.2
numpy>=1.26
redis>=5.0
django-redis>=5.4
python-dotenv>=1.0
